refactor!: move OAuth and token auth code to auth.py
This doesn't belong in api.py
